backend/requirements.txt
fastapi=0.104.0
uvicorn[standard]=0.24.0
python-multipart=0.0.6
celery=5.3.0
redis=5.0.0
google-generativeai=0.3.0
groq=0.4.0
faster-whisper=1.0.0
httpx=0.25.0
python-dotenv=1.0.0
pydantic=2.0.0
openai
